rename the streaming examples
